requirements.txt
google-generativeai
google-adk
yfinance
ta
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
python-multipart
fastmcp
ddgs




